internal/contour: move translatorCache to its own file
Move the translatorCache outside the monolithic translator.go.
